refactor: centralize single-store helpers shared by promotions and payment methods
Consolidate the two near-identical LegacyMultiStoreSupport concerns into one
shared Spree::LegacyMultiStoreSupport (base-class-aware deprecation messages),
and move assign_default_store into Spree::SingleStoreResource as an opt-in
assign_default_store_on_create macro. Adds the missing payment-method
idempotency test and documents the dependent: :nullify choice.
