+++ spree/core/app/models/spree/promotion.rb
@@ -9,7 +9,7 @@ module Spree
       include Spree::Security::Promotions
     end
     # Multi-store sharing moved to the spree_multi_store extension in 5.6.
-    include Spree::Promotion::LegacyMultiStoreSupport unless defined?(SpreeMultiStore)
+    include Spree::LegacyMultiStoreSupport unless defined?(SpreeMultiStore)
 
     publishes_lifecycle_events
 
@@ -40,13 +40,13 @@ module Spree
     has_many :order_promotions, class_name: 'Spree::OrderPromotion'
     has_many :orders, through: :order_promotions, class_name: 'Spree::Order'
     belongs_to :store, class_name: 'Spree::Store', optional: true
+    assign_default_store_on_create
 
     after_save :apply_pending_rules_and_actions, if: :pending_rules_or_actions?
 
     #
     # Callbacks
     #
-    before_validation :assign_default_store, on: :create, if: -> { store.nil? }
     before_validation :set_code_to_nil, if: -> { multi_codes? || automatic? }
     before_validation :set_number_of_codes_to_nil, if: -> { automatic? || !multi_codes? }
     before_validation :set_usage_limit_to_nil, if: -> { multi_codes? }
@@ -302,10 +302,6 @@ module Spree
     end
 
     private
-
-    def assign_default_store
-      self.store ||= Spree::Current.store || Spree::Store.default
-    end
 
     def apply_pending_rules_and_actions
       flush_pending_typed_association(:promotion_rules)
